Tee 4.21, testejä rikki

+++ osa_4/controllers/blogs.js
@@ -45,16 +45,26 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
-  /*
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json(
+      { error: 'token missing or invalid' })
+  }
+
   const blog = await Blog.findById(request.params.id)
-  if (blog) {
-    await Blog.findByIdAndRemove(blog.id)
+
+  const user = await User.findById(decodedToken.id)
+
+  if (!blog) {
+    response.status(400).json({ error : 'blog not found' })
+  } else if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
   } else {
-    response.status(404).end()
-  }*/
+    response.status(401).json(
+      { error: 'user not authorized to delete this blog' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
